# docs-site/openapi/public.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Feedling API",
    "version": "v1",
    "description": "HTTP API for Feedling accounts, encrypted chat, agent memory, model routing, perception, and proactive experiences."
  },
  "servers": [
    {
      "url": "https://api.feedling.app",
      "description": "Production"
    },
    {
      "url": "https://test-api.feedling.app",
      "description": "Staging"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "RuntimeTokenAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/chat/history": {
      "get": {
        "summary": "Chat History",
        "operationId": "chat_history_v1_chat_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum messages in this page.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 200
            },
            "example": 50
          },
          {
            "name": "after_seq",
            "in": "query",
            "required": false,
            "description": "Lossless forward cursor. Return messages whose durable append sequence is greater than this value; takes precedence over timestamp cursors.",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "example": 1042
          },
          {
            "name": "before_seq",
            "in": "query",
            "required": false,
            "description": "Lossless backward cursor. Return messages whose durable append sequence is less than this value; takes precedence over timestamp cursors.",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1042
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "description": "Return messages with ts strictly greater than this watermark.",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "example": 1783962000.0
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "Return older messages with ts strictly less than this watermark; takes precedence over since.",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "example": 1783962000.0
          },
          {
            "name": "include_image_body",
            "in": "query",
            "required": false,
            "description": "Set false to omit image and oversized inline bodies.",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "example": false
          },
          {
            "name": "include_image_bodies",
            "in": "query",
            "required": false,
            "description": "Compatibility alias for include_image_body.",
            "schema": {
              "type": "boolean"
            },
            "deprecated": true
          }
        ],
        "x-feedling-contract-level": "documented",
        "description": "Read encrypted chat history. Use oldest_seq as before_seq for lossless older paging and latest_seq as after_seq for lossless forward paging; timestamp watermarks remain for compatibility."
      },
      "delete": {
        "summary": "Chat History Clear",
        "operationId": "chat_history_clear_v1_chat_history_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatHistoryClearRequest"
              }
            }
          }
        },
        "x-feedling-contract-level": "documented"
      }
    },
    "/v1/chat/message": {
      "post": {
        "summary": "Chat Message",
        "operationId": "chat_message_v1_chat_message_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatTransportRequest"
              }
            }
          }
        },
        "x-feedling-contract-level": "documented",
        "description": "Store a user chat message as a v1 ciphertext envelope; the server never decrypts it. If the envelope carries a content_pk_fpr label that does not match the user's currently registered content key, the write is rejected with 409 content_pk_fpr_mismatch (re-fetch whoami and re-seal); unlabeled envelopes are accepted for compatibility."
      }
    },
    "/v1/chat/messages/{message_id}/body": {
      "get": {
        "summary": "Chat Message Body",
        "operationId": "chat_message_body_v1_chat_messages__message_id__body_get",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path or request parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "x-feedling-contract-level": "documented"
      }
    },
    "/v1/chat/poll": {
      "get": {
        "summary": "Chat Poll",
        "operationId": "chat_poll_v1_chat_poll_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "description": "Return messages newer than this Unix timestamp.",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "example": 1783962000.0
          },
          {
            "name": "timeout",
            "in": "query",
            "required": false,
            "description": "Long-poll wait in seconds; the server caps it at 60.",
            "schema": {
              "type": "number",
              "minimum": 0,
              "maximum": 60,
              "default": 30
            },
            "example": 30
          },
          {
            "name": "consumer_id",
            "in": "query",
            "required": false,
            "description": "Stable resident consumer identifier.",
            "schema": {
              "type": "string",
              "maxLength": 160
            },
            "example": "resident_mbp_01"
          },
          {
            "name": "claim",
            "in": "query",
            "required": false,
            "description": "Whether this poll may claim queued work.",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "example": true
          },
          {
            "name": "X-Feedling-Consumer",
            "in": "header",
            "required": false,
            "description": "Resident consumer name. Used when X-Feedling-Consumer-Id is absent.",
            "schema": {
              "type": "string",
              "maxLength": 160
            },
            "example": "feedling-resident"
          },
          {
            "name": "X-Feedling-Consumer-Id",
            "in": "header",
            "required": false,
            "description": "Stable resident consumer identifier.",
            "schema": {
              "type": "string",
              "maxLength": 160
            },
            "example": "resident_mbp_01"
          },
          {
            "name": "X-Feedling-Consumer-Version",
            "in": "header",
            "required": false,
            "description": "Resident consumer release version.",
            "schema": {
              "type": "string",
              "maxLength": 80
            },
            "example": "1.4.0"
          },
          {
            "name": "X-Feedling-Consumer-Commit",
            "in": "header",
            "required": false,
            "description": "Resident consumer source revision.",
            "schema": {
              "type": "string",
              "maxLength": 80
            },
            "example": "a1b2c3d"
          },
          {
            "name": "X-Feedling-Consumer-Capabilities",
            "in": "header",
            "required": false,
            "description": "Comma-separated capabilities advertised by the current official resident poll.",
            "schema": {
              "type": "string",
              "maxLength": 500
            },
            "example": "vision_observer_v1"
          },
          {
            "name": "X-Feedling-Consumer-Compat-Commit",
            "in": "header",
            "required": false,
            "description": "Optional poll-only backend target that an older running resident intentionally skipped as irrelevant. It suppresses commit-mismatch maintenance only when a running commit is present and this value matches the backend's current expected commit.",
            "schema": {
              "type": "string",
              "maxLength": 80
            },
            "example": "a1b2c3d"
          },
          {
            "name": "X-Feedling-Decrypt-Status",
            "in": "header",
            "required": false,
            "description": "Current resident decrypt-source health. Report this on poll heartbeats; omitting it on a later official poll clears the previous report to unknown.",
            "schema": {
              "type": "string",
              "enum": [
                "ok",
                "degraded",
                "unconfigured",
                "unreachable"
              ]
            },
            "example": "ok"
          },
          {
            "name": "X-Feedling-Decrypt-Checked-At",
            "in": "header",
            "required": false,
            "description": "Unix epoch seconds when the resident last confirmed the reported decrypt status. The backend treats missing, invalid, future, or stale values as unknown.",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "example": 1784625600.125
          }
        ],
        "x-feedling-contract-level": "documented",
        "description": "Long-poll and optionally claim resident chat work. Official residents report their running commit and may report an intentionally skipped compatible backend target with X-Feedling-Consumer-Compat-Commit. They also report decrypt-source status and its confirmation time on every poll heartbeat with X-Feedling-Decrypt-Status and X-Feedling-Decrypt-Checked-At."
      }
    },
    "/v1/chat/response": {
      "post": {
        "summary": "Chat Response",
        "operationId": "chat_response_v1_chat_response_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "parameters": [
          {
            "name": "consumer_id",
            "in": "query",
            "required": false,
            "description": "Stable resident consumer identifier.",
            "schema": {
              "type": "string",
              "maxLength": 160
            },
            "example": "resident_mbp_01"
          },
          {
            "name": "X-Feedling-Consumer",
            "in": "header",
            "required": false,
            "description": "Resident consumer name. Used when X-Feedling-Consumer-Id is absent.",
            "schema": {
              "type": "string",
              "maxLength": 160
            },
            "example": "feedling-resident"
          },
          {
            "name": "X-Feedling-Consumer-Id",
            "in": "header",
            "required": false,
            "description": "Stable resident consumer identifier.",
            "schema": {
              "type": "string",
              "maxLength": 160
            },
            "example": "resident_mbp_01"
          },
          {
            "name": "X-Feedling-Consumer-Version",
            "in": "header",
            "required": false,
            "description": "Resident consumer release version.",
            "schema": {
              "type": "string",
              "maxLength": 80
            },
            "example": "1.4.0"
          },
          {
            "name": "X-Feedling-Consumer-Commit",
            "in": "header",
            "required": false,
            "description": "Resident consumer source revision.",
            "schema": {
              "type": "string",
              "maxLength": 80
            },
            "example": "a1b2c3d"
          },
          {
            "name": "X-Feedling-Consumer-Capabilities",
            "in": "header",
            "required": false,
            "description": "Comma-separated capabilities advertised by the current official resident poll.",
            "schema": {
              "type": "string",
              "maxLength": 500
            },
            "example": "vision_observer_v1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatResponseRequest"
              }
            }
          }
        },
        "x-feedling-contract-level": "documented",
        "description": "Store an agent reply as a v1 ciphertext envelope (plus optional thinking envelope and encrypted file_followups). A text primary and its file cards commit as one ordered transaction. Replies carrying reply_to_message_id are finalized atomically across backend workers: exactly one request inserts the reply and marks the parent answered, while a losing contender returns 409 already_answered without storing its reply. A hidden source=verify_ping reply is accepted only when reply_to_message_id identifies an outstanding verify ping exactly. role=system notices bypass reply exclusivity. Labeled envelopes sealed to a key that is no longer the user's registered content key are rejected with 409 content_pk_fpr_mismatch — the writer should re-fetch whoami, re-seal, and retry once."
      }
    },
    "/v1/chat/turn-activity/{turn_id}": {
      "get": {
        "summary": "Chat Turn Activity",
        "operationId": "chat_turn_activity_v1_chat_turn_activity__turn_id__get",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Turn Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path or request parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "x-feedling-contract-level": "documented",
        "description": "Read display-safe activity for one V1 resident or Runtime V2 chat turn. V2 events come from backend jobs and tool dispatch; V1 events come from the authenticated resident io_cli boundary and are durably scoped to an existing user message. Both runtimes expose only bounded identifiers, state, timing, and result classification. Successful memory_search/memory_fetch events include the confirmed returned-item count and, only when every item uses the canonical bucket taxonomy, a complete category-count breakdown. Tool arguments, result bodies, assistant prose, reasoning, and custom bucket labels are never returned."
      }
    },
    "/v1/chat/turn-activity/{turn_id}/events": {
      "post": {
        "summary": "Chat Turn Activity Event",
        "operationId": "chat_turn_activity_event_v1_chat_turn_activity__turn_id__events_post",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Turn Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path or request parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatActivityEventRequest"
              }
            }
          }
        },
        "x-feedling-contract-level": "documented",
        "description": "Append one authenticated V1 resident tool transition. This endpoint is used by the shipped resident io_cli runtime, accepts only running/success/failure plus display-safe fixed metadata, rejects V2-owned users, and never accepts tool arguments, model prose, or result bodies."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenericJsonResponse": {
        "type": "object",
        "description": "JSON response object. Endpoint-specific fields may be added compatibly.",
        "additionalProperties": true
      },
...
      "ChatHistoryClearRequest": {
        "type": "object",
        "required": [
          "confirm"
        ],
        "properties": {
          "confirm": {
            "type": "string",
            "const": "clear-chat-history"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Long-lived Feedling user API key."
      },
      "RuntimeTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Feedling-Runtime-Token",
        "description": "Short-lived scoped token used by hosted runtimes."
      },
      "RelayTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Relay-Token",
        "description": "Notify-relay token (nrt_…) issued by POST /v1/notify-relay/register. Not an account credential."
      }
    }
  }
}